feat: Simplified the code and adjusted the functions
