+++ packages/checker/checking_wahiddl.py
@@ -1,10 +1,10 @@
 import os
-from packages.core import system_os
+from packages.core import version_info
 wahiddl_folder_name = "wahid-dl"
 def checking_wahiddl_folder_existed():
-    if system_os.get_system_os() == "Windows":
+    if version_info.vi_platform() == "Windows":
         path = os.path.join("C:\\", "wahid-dl")
-    elif system_os.get_system_os() == "Linux":
+    elif version_info.vi_platform() == "Linux":
         path = os.path.expanduser("~/wahid-dl")
     else:
         return False, "Unsupported OS"
