--- packages/core/system_os.py
@@ -1,11 +1,0 @@
-import sys
-def get_system_os():
-    if sys.platform  == "win32":
-        system_os = "Windows"
-        return system_os
-    elif sys.platform  == "linux":
-        system_os = "Linux"
-        return system_os
-    elif sys.platform == "darwin":
-        system_os = "macOS"
-        return system_os
