+++ packages/core/ui.py
@@ -1,5 +1,5 @@
-from packages.core import system_os
 from packages.core import version_info
+#ASCII art
 def ASCII_art():
     return("      ###       ###        ###         ###     ###         #######      ########             #########       ###\n"
            "     ###       ###      ###   ##      ###     ###         ###          ###    ###           ###    ###      ###\n"
@@ -8,38 +8,39 @@ def ASCII_art():
            "  ### ##  # ###    ###      ##     ###     ###         ###          ###    ###           ###    ###      ###\n"
            " #####    ###     ###      ##     ###     ###         ###          ###    ###           ###    ###      ###\n"
            "####      ##     ###      ##     ###     ###     #######          #########            #########       ###")
-system_os = system_os.get_system_os()
-programming_language = version_info.version_programming_language()
-def get_title_wahiddl():
-    title_wahiddl = "wahid-dl [" + system_os + "] [" + programming_language + "]"+ "\n" + version_info.version_outline()
+#Title with version info
+vi=str(" [" + str(version_info.vi_platform()) + "] [" + str(version_info.vi_detail()[0]) + "]v" + str(version_info.vi_detail()[1] + version_info.vi_detail()[2]) + "-" + str(version_info.vi_detail()[3]))
+def ui_title_wahiddl():
+    title_wahiddl = "wahid-dl" + vi
     return title_wahiddl
-def get_title_wahiddl_tool():
-    title_wahiddl_tool = "wahid-dl Tool [" + system_os + "] [" + programming_language + "]"+ "\n" + version_info.version_outline()
+def ui_title_wahiddl_tool():
+    title_wahiddl_tool = "wahid-dl Tool" + vi
     return title_wahiddl_tool
-def get_title_wahiddl_devmode():
-    title_wahiddl_devmode = "wahid-dl Dev Mode [" + system_os + "] [" + programming_language + "]" + "\n" + version_info.version_outline()
+def ui_title_wahiddl_devmode():
+    title_wahiddl_devmode = "wahid-dl Dev Mode" + vi
     return title_wahiddl_devmode
-def get_title_wahiddl_format_checking_tool():
-    title_wahiddl_format_checking_tool = "wahid-dl Format Checking Tool [" + system_os + "] [" + programming_language + "]" + "\n" + version_info.version_outline()
+def ui_title_wahiddl_format_checking_tool():
+    title_wahiddl_format_checking_tool = "wahid-dl Format Checking Tool" + vi
     return title_wahiddl_format_checking_tool
-def get_title_wahiddl_S():
-    title_wahiddl_S = "wahid-dl for wahid-dl for Subtitle [" + system_os + "] [" + programming_language + "]" + "\n" + version_info.version_outline()
+def ui_title_wahiddl_S():
+    title_wahiddl_S = "wahid-dl for wahid-dl for Subtitle" + vi
     return title_wahiddl_S
-def get_title_wahiddl_QS():
-    title_wahiddl_QS = "wahid-dl for Quality Selection [" + system_os + "] [" + programming_language + "]" + "\n" + version_info.version_outline()
+def ui_title_wahiddl_QS():
+    title_wahiddl_QS = "wahid-dl for Quality Selection" + vi
     return title_wahiddl_QS
-def get_title_wahiddl_live():
-    title_wahiddl_live = "wahid-dl for Live [" + system_os + "] [" + programming_language + "]" + "\n" + version_info.version_outline()
+def ui_title_wahiddl_live():
+    title_wahiddl_live = "wahid-dl for Live" + vi
     return title_wahiddl_live
-def get_title_wahiddl_list():
-    title_wahiddl_list = "wahid-dl for List [" + system_os + "] [" + programming_language + "]" + "\n" + version_info.version_outline()
+def ui_title_wahiddl_list():
+    title_wahiddl_list = "wahid-dl for List" + vi
     return title_wahiddl_list
-def get_title_wahiddl_audio():
-    title_wahiddl_audio = "wahid-dl for Audio [" + system_os + "] [" + programming_language + "]" + "\n" + version_info.version_outline()
+def ui_title_wahiddl_audio():
+    title_wahiddl_audio = "wahid-dl for Audio" + vi
     return title_wahiddl_audio
-def get_title_wahiddl_CS():
-    title_wahiddl_CS = "wahid-dl for Cookies Support [" + system_os + "] [" + programming_language + "]" + "\n" + version_info.version_outline()
+def ui_title_wahiddl_CS():
+    title_wahiddl_CS = "wahid-dl for Cookies Support" + vi
     return title_wahiddl_CS
+#Unified UI for task execution status
 def ui_start():
     print ("------------------------------------------------------------")
     print ("任務開始執行")
