+++ wahid-dl for List.py
@@ -4,7 +4,7 @@ from packages.core import ui
 
 if sys.platform == "win32":
     print (ui.ASCII_art())
-    print (ui.get_title_wahiddl_list())
+    print (ui.ui_title_wahiddl_list())
     print ("------------------------------------------------------------")
     mode = ""
     while mode != "0":
@@ -57,7 +57,7 @@ if sys.platform == "win32":
         break
 elif sys.platform == "linux":
     print (ui.ASCII_art())
-    print (ui.get_title_wahiddl_list())
+    print (ui.ui_title_wahiddl_list())
     print ("------------------------------------------------------------")
     mode = ""
     while mode != "0":
