+++ wahid-dl for Subtitle.py
@@ -4,7 +4,7 @@ from packages.core import ui
 
 if sys.platform == "win32":
     print (ui.ASCII_art())
-    print (ui.get_title_wahiddl_S())
+    print (ui.ui_title_wahiddl_S())
     print ("------------------------------------------------------------")
     mode = ""
     while mode != "0":
@@ -42,7 +42,7 @@ if sys.platform == "win32":
         break
 elif sys.platform == "linux":
     print (ui.ASCII_art())
-    print (ui.get_title_wahiddl_S())
+    print (ui.ui_title_wahiddl_S())
     print ("------------------------------------------------------------")
     mode = ""
     while mode != "0":
